Riccardo 04 02 2024 update review

+++ server/models/model.js
@@ -141,10 +141,29 @@ async function fetchReviewsById(book_id) {
 async function removeReviewById(review_id) {
   try {
     const findReview = await Review.findById(review_id);
+    if (!findReview) {
+      return Promise.reject({ status: 404, msg: "Review not found"})
+    }
     const bookId = findReview.bookId;
     const deletedReview = await Review.findByIdAndDelete(review_id);
     await updateBookRating(bookId);
     return deletedReview;
+  } catch (error) {
+    console.log(error);
+  }
+}
+
+async function amendReviewById(review_id, reviewBody, rating) {
+  try {
+    const findReview = await Review.findById(review_id);
+    if (!findReview) {
+      return Promise.reject({ status: 404, msg: "Review not found"});
+    }
+    const bookId = findReview.bookId;
+    const newUpdates = { reviewBody, rating };
+    const updatedReview = await Review.findByIdAndUpdate(review_id, newUpdates, {new : true});
+    await updateBookRating(bookId);
+    return updatedReview;
   } catch (error) {
     console.log(error);
   }
@@ -157,4 +176,5 @@ module.exports = {
   sendBookReview,
   fetchReviewsById,
   removeReviewById,
+  amendReviewById
 };
